module name refactored to ls-cosmos
