+++ proto/lscosmos/v1beta1/packet.proto
@@ -3,7 +3,7 @@ package lscosmos.v1beta1;
 
 // this line is used by starport scaffolding # proto/packet/import
 
-option go_package = "github.com/persistenceOne/pstake-native/x/lscosmos/types";
+option go_package = "github.com/persistenceOne/pstake-native/x/ls-cosmos/types";
 
 message LscosmosPacketData {
     oneof packet {
