+++ proto/lscosmos/v1beta1/params.proto
@@ -3,7 +3,7 @@ package lscosmos.v1beta1;
 
 import "gogoproto/gogo.proto";
 
-option go_package = "github.com/persistenceOne/pstake-native/x/lscosmos/types";
+option go_package = "github.com/persistenceOne/pstake-native/x/ls-cosmos/types";
 
 // Params defines the parameters for the module.
 message Params {
