+++ x/ls-cosmos/handler.go
@@ -1,12 +1,12 @@
-package lscosmos
+package ls_cosmos
 
 import (
 	"fmt"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/persistenceOne/pstake-native/x/lscosmos/keeper"
-	"github.com/persistenceOne/pstake-native/x/lscosmos/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	"github.com/persistenceOne/pstake-native/x/ls-cosmos/keeper"
+	"github.com/persistenceOne/pstake-native/x/ls-cosmos/types"
 )
 
 // NewHandler ...
