+++ x/ls-cosmos/keeper/grpc_query.go
@@ -1,7 +1,7 @@
 package keeper
 
 import (
-	"github.com/persistenceOne/pstake-native/x/lscosmos/types"
+	"github.com/persistenceOne/pstake-native/x/ls-cosmos/types"
 )
 
 var _ types.QueryServer = Keeper{}
