+++ x/ls-cosmos/keeper/msg_server.go
@@ -1,7 +1,7 @@
 package keeper
 
 import (
-	"github.com/persistenceOne/pstake-native/x/lscosmos/types"
+	"github.com/persistenceOne/pstake-native/x/ls-cosmos/types"
 )
 
 type msgServer struct {
