+++ x/ls-cosmos/keeper/params.go
@@ -1,18 +1,16 @@
 package keeper
 
 import (
-	"github.com/persistenceOne/pstake-native/x/lscosmos/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/persistenceOne/pstake-native/x/ls-cosmos/types"
 )
 
 // GetParams get all parameters as types.Params
 func (k Keeper) GetParams(ctx sdk.Context) types.Params {
-	return types.NewParams(
-	)
+	return types.NewParams()
 }
 
 // SetParams set the params
 func (k Keeper) SetParams(ctx sdk.Context, params types.Params) {
 	k.paramstore.SetParamSet(ctx, &params)
 }
-
