+++ x/ls-cosmos/types/errors.go
