--- x/lscosmos/genesis.go
@@ -1,36 +1,0 @@
-package lscosmos
-
-import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/persistenceOne/pstake-native/x/lscosmos/keeper"
-	"github.com/persistenceOne/pstake-native/x/lscosmos/types"
-)
-
-// InitGenesis initializes the capability module's state from a provided genesis
-// state.
-func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) {
-    // this line is used by starport scaffolding # genesis/module/init
-k.SetPort(ctx, genState.PortId)
-// Only try to bind to port if it is not already bound, since we may already own
-// port capability from capability InitGenesis
-if !k.IsBound(ctx, genState.PortId) {
-	// module binds to the port on InitChain
-	// and claims the returned capability
-	err := k.BindPort(ctx, genState.PortId)
-	if err != nil {
-		panic("could not claim port capability: " + err.Error())
-	}
-}
-	k.SetParams(ctx, genState.Params)
-}
-
-// ExportGenesis returns the capability module's exported genesis.
-func ExportGenesis(ctx sdk.Context, k keeper.Keeper) *types.GenesisState {
-	genesis := types.DefaultGenesis()
-	genesis.Params = k.GetParams(ctx)
-
-    genesis.PortId = k.GetPort(ctx)
-// this line is used by starport scaffolding # genesis/module/export
-
-    return genesis
-}
