--- x/lscosmos/types/genesis_test.go
@@ -1,40 +1,0 @@
-package types_test
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-	"github.com/persistenceOne/pstake-native/x/lscosmos/types"
-)
-
-func TestGenesisState_Validate(t *testing.T) {
-    for _, tc := range []struct {
-    		desc          string
-    		genState      *types.GenesisState
-    		valid bool
-    } {
-        {
-            desc:     "default is valid",
-            genState: types.DefaultGenesis(),
-            valid:    true,
-        },
-        {
-            desc:     "valid genesis state",
-            genState: &types.GenesisState{
-            	PortId: types.PortID,
-                // this line is used by starport scaffolding # types/genesis/validField
-            },
-            valid:    true,
-        },
-        // this line is used by starport scaffolding # types/genesis/testcase
-    } {
-        t.Run(tc.desc, func(t *testing.T) {
-            err := tc.genState.Validate()
-            if tc.valid {
-                require.NoError(t, err)
-            } else {
-                require.Error(t, err)
-            }
-        })
-    }
-}
